Make the braintree payment form bullet proof
Previously, the tests failed on fast computers due to the fact
poltergeist could submit the form faster than the javascript could be
ready to accept input.

We Deface the frontend here (this is bad, but temporary hopefully) to
fix this issue. A permanent fix would be adding a system to frontend to
handle third party payment providers properly. This button should always
be disabled until the interface is ready to accept input.